elearning_platform/quiz_service/requirements.txt
asgiref==3.11.1
certifi==2026.2.25
charset-normalizer==3.4.6
Django==4.2
django-cors-headers==4.3.0
djangorestframework==3.14.0
idna==3.11
pika==1.3.2
psycopg2-binary==2.9.11
PyJWT==2.8.0
pytz==2026.1.post1
requests==2.31.0
sqlparse==0.5.0
tzdata==2025.3
urllib3==2.6.3
python-consul==1.1.0